requirements.txt
# auto-gptq==0.4.2+cu117
boto3==1.28.58
fastparquet==2023.8.0
gradio==3.46.0
optimum==1.13.2
pandas==2.0.3
peft==0.5.0
accelerate==0.23.0
python-dotenv==1.0.0
sagemaker==2.189.0
tqdm==4.66.1
transformers==4.34.0
trl==0.7.1
datasets==2.14.5
fire==0.5.0